Beta test for 7speaking.user.js enhancement
Changelogs :
NOTE : CODE IS NOT WORKING YET

- General Optimizations :
 - Added the await keyword before each recursive call to the routes() function. This ensures that each call to routes() is completed before moving on to the next call.

 - waitForQuerySelector Function :
Replaced setInterval with a while loop with await wait(interval). This avoids creating a new interval at each iteration.

- getReactElement Function :
Used Object.values(e).find(...) instead of the for loop. This makes the code more readable and concise.

- getInputElement Function :
Used the logical OR operator (||) to simplify the search for the input element. This avoids code duplication.
Replaced the for loop with the find method of the Array object to find the button corresponding to the answer. This makes the code more readable and concise.

- completeExam Function :
Replaced for...in loops with forEach methods of the Array object to iterate over rows and columns. This makes the code more readable and concise.

Used the Object.entries method to iterate over the properties of the answer object. This allows access to both the key and value of each property, simplifying the code.

- routes Function :
Added the await keyword before each recursive call to the routes() function. This ensures that each call to routes() is completed before moving on to the next call. This can help avoid potential issues related to asynchronous execution.


These modifications aim to improve code readability, reduce code duplication, and optimize asynchronous execution.